third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/xmlhttprequest/open-with-invalid-argument-is-noop-expected.txt

This is a testharness.js-based test.
PASS open() with an invalid method is no-op 
PASS open() with a forbidden method is no-op 
PASS open() with an invalid URL is no-op 
Harness: the test ran to completion.
